pyproject.toml
[project]
name = "proyecto-timely-response"
version = "0.1.0"
description = "Timely response classification + business metrics"
readme = "README.md"
requires-python = ">=3.12"
authors = [{ name = "Abel_E" }]

dependencies = [
    "pandas>=2.0",
    "numpy>=1.25",
    "scikit-learn>=1.2",
    "matplotlib>=3.7",
    "joblib>=1.3",
    "pyyaml>=6.0",
    "streamlit>=1.33",
    "python-pptx>=0.6",
    "accelerate>=1.8.1",
    "azure-ai-inference>=1.0.0b9",
    "comet-ml>=3.49.12",
    "diffusers>=0.34.0",
    "evidently>=0.7.11",
    "gradio>=5.38.0",
    "gymnasium[classic-control,other]>=1.2.0",
    "huggingface-hub>=0.33.4",
    "hyperopt>=0.2.7",
    "ipykernel>=6.29.5",
    "langchain-google-genai>=2.1.8",
    "langchain-openai>=0.3.28",
    "matplotlib>=3.10.3",
    "mlflow>=3.1.1",
    "nbformat>5",
    "nest-asyncio>=1.6.0",
    "openai>=1.95.1",
    "pydantic-ai>=0.4.3",
    "pydot>=4.0.1",
    "python-dotenv>=1.1.1",
    "skrub>=0.6.0",
    "tensorflow>=2.19.0",
    "tf-keras>=2.19.0",
    "transformers>=4.53.2",
    "virtualenv>=20.32.0",
    "scipy",
    "statsmodels",
    "seaborn==0.13.2",
    "black>=23.7.0",
    "flake8>=6.0.0",
    "jupyterlab>=4.0",
    "ipykernel>=6.0",
    "watchdog>=3.0",
]


[project.scripts]
train-timely = "src.train_cli:main"
predict-timely = "src.predict_cli:main"

[tool.black]
line-length = 120
